Separate CSA2 layer ownership policy from shared attention computation
